Skill user passed in as method parameter instead of stored in field.
The Skill class represents a specification for a skill. Actual usage
of the skill is represented by the new SkillUsage class, made up of a
skill, its user, and its target.

The previous pattern of storing a CharacterType turned out to be very
brittle, indicating to me that I was using a wrong abstraction. This
way gives more consistent access to Character data, as before the recent
CharacterType change, while still reducing the reliance on copying
prototypes.